reorganize core library extensions
