+++ lib/asciidoctor/core_ext.rb
@@ -1,8 +1,8 @@
-require 'asciidoctor/core_ext/object/nil_or_empty'
-require 'asciidoctor/core_ext/string/limit'
-unless RUBY_ENGINE == 'opal'
-  unless RUBY_MIN_VERSION_1_9
-    require 'asciidoctor/core_ext/string/chr'
-    require 'asciidoctor/core_ext/symbol/length'
-  end
+require 'asciidoctor/core_ext/nil_or_empty'
+if RUBY_MIN_VERSION_1_9
+  require 'asciidoctor/core_ext/string/limit'
+elsif RUBY_ENGINE != 'opal'
+  require 'asciidoctor/core_ext/1.8.7/string/chr'
+  require 'asciidoctor/core_ext/1.8.7/string/limit'
+  require 'asciidoctor/core_ext/1.8.7/symbol/length'
 end
