+++ lib/asciidoctor/core_ext/1.8.7/string/chr.rb
