+++ lib/asciidoctor/core_ext/1.8.7/symbol/length.rb
