+++ lib/asciidoctor/core_ext/nil_or_empty.rb
@@ -1,6 +1,6 @@
 # A core library extension that defines the method nil_or_empty? as an alias to
 # optimize checks for nil? or empty? on common object types such as NilClass,
-# String, Array and Hash.
+# String, Array, Hash, and Numeric.
 
 class NilClass
   alias :nil_or_empty? :nil? unless method_defined? :nil_or_empty?
